backend/pyproject.toml
[project]
name = "therapyai_backend"
version = "0.1.0"
description = "Backend for therapyai"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "supabase>=2.27.0",
    "uvicorn[standard]>=0.35.0",
]
